EverySingleStreet

Documentation for EverySingleStreet.

EverySingleStreet.StreetSegment

EverySingleStreet.calculate_streetpath

EverySingleStreet.download

EverySingleStreet.filter_candidates!

EverySingleStreet.filter_path

EverySingleStreet.get_candidate_on_way

EverySingleStreet.get_candidate_probability

EverySingleStreet.get_candidates

EverySingleStreet.get_candidates_from_idx

EverySingleStreet.get_interpolation_point

EverySingleStreet.get_interpolation_val

EverySingleStreet.get_lla

EverySingleStreet.get_matching_candidates

EverySingleStreet.get_next_node_id

EverySingleStreet.get_prev_node_id

EverySingleStreet.get_segments

EverySingleStreet.get_way_kdtree

EverySingleStreet.getxy_from_lat_lon

EverySingleStreet.map_path

EverySingleStreet.parse_map

EverySingleStreet.point_linesegment_distance

EverySingleStreet.shortest_candidate_path

EverySingleStreet.shortest_path_distance

EverySingleStreet.transition_probability

EverySingleStreet.StreetSegment — Type

struct StreetSegment

A street segment contains of two candidates which have the following property: They are both on the same way and have the same direction.

EverySingleStreet.calculate_streetpath — Method

calculate_streetpath(ame, subpath_id, candidates, city_map)

Generate a StreetPath from a list of candidates obtained by map_path.

EverySingleStreet.download — Method

download(place_name, filepath)

Download the road network of the given place and write it as a json file to the given filepath

EverySingleStreet.filter_candidates! — Method

filter_candidates!(candidates; closer_dist=25)

Filter out candidates which are further away than closer_dist if there is at least candidate which is closer than closer_dist.

EverySingleStreet.filter_path — Method

filter_path(gps_points, dist)

Return a new path for which the minimum distance between two consecutive points is at least dist.

EverySingleStreet.get_candidate_on_way — Method

get_candidate_on_way(city_map, p, way::Way, trans, rev_trans; rev=false)

Get the best candidate of point p on the given way. trans and rev_trans are transformations mapping from LLA to x,y and back. rev can be set to true to reverse the direction of way.

EverySingleStreet.get_candidate_probability — Method

get_candidate_probability(candidate::Candidate; sigma=10)

Return the emission probability of the given candidate and the standard deviation of the gps error.

EverySingleStreet.get_candidates — Method

get_candidates(map, path)

Get a list of Candidate for each gps point in path given the underlying city_map

EverySingleStreet.get_candidates_from_idx — Method

get_candidates_from_idx(vec_candidates, candidate_idxs)

Return candidates from an nested vector of candiates like [[c1,c2],[c3]] and candiate_idxs which are 1d like [1,3] would return [c1, c3].

EverySingleStreet.get_interpolation_point — Method

get_interpolation_point(a, b, t::Float64)

Get the value of a+t*(b-a)

EverySingleStreet.get_interpolation_val — Method

get_interpolation_val(p::Point, a::Point, b::Point)

Get an interpolation value of point p between a line segment from a to b where a+t*(b-a) describes the point closes to p. Return p which is between 0 and 1.

EverySingleStreet.get_lla — Method

get_lla(p, trans)

Get a LLA formatted position from a point and a Geodesy transformation.

EverySingleStreet.get_matching_candidates — Method

get_matching_candidates(city_map, way_ids, p, origin_lla; maximum_dist=30)

Get matching candidates for a given point p and a list of possible way ids. Only return candidates which have a maximum_dist (in m) to the way.

EverySingleStreet.get_next_node_id — Method

get_next_node_id(candidate::Candidate)

Get the next node id given a candidate. This depends on way_is_reverse of the candidate. Can be used to create a StreetSegment.

EverySingleStreet.get_prev_node_id — Method

get_prev_node_id(candidate::Candidate)

Get the previous node id given a candidate. This depends on way_is_reverse of the candidate. Can be used to create a StreetSegment.

EverySingleStreet.get_segments — Method

get_segments(city_map, current_candidate, next_candidate, sp)

Given two candidates which can't be directly connected and a list of ids which form the shortest path between those candidates this function returns a list of StreetSegment which connect the two candidates.

EverySingleStreet.get_way_kdtree — Method

get_way_kdtree(city_map)

Create a KDTree for all ways in the given city map. Return

a mapping from id to the way id

the kd tree

The radius for a reasonable in range search (51% of the longest line between two nodes)

EverySingleStreet.getxy_from_lat_lon — Method

getxy_from_lat_lon(lat, lon, trans)

Return x, y position coordinates from lat and lon given a Geodesy trensformation.

EverySingleStreet.map_path — Method

map_path(city_map, gpxfile)

Map a path of gps points to the best matching candidates for the path.

Filter out points on the path which are closer together than 25m

Compute candidates for each point in the remaining path

Compute emission probabilties

Compute transition_probabilities

Use the Viterbi algorithm to compute the most likely path of the candidates

Return a list of lists of candidates as a path might not continously have candidates. Then it's split up into several parts.

EverySingleStreet.parse_map — Method

parse_map(fpath)

Return a Map object from the given json file path that was created using the download function.

EverySingleStreet.point_linesegment_distance — Method

point_linesegment_distance(p::Point, a::Point, b::Point)

Get the smallest distance between point p and the line segment from a to b.

EverySingleStreet.shortest_candidate_path — Method

shortest_candidate_path(from::Candidate, to::Candidate, city_map)

Return the shortest path as node ids from one candidate to another given a city map.

EverySingleStreet.shortest_path_distance — Method

shortest_path_distance(from::Candidate, to::Candidate, city_map)

Return the shortest path distance based on equation (9) in the fmm paper

EverySingleStreet.transition_probability — Method

transition_probability(from::Candidate, to::Candidate, city_map)

Return the transitionprobability based on equation 10 in the [fmm paper](https://www.tandfonline.com/doi/pdf/10.1080/13658816.2017.1400548?casatoken=RrxNeRXfRfkAAAAA:2IA6z4Pu-tKEcSaK44AUgQxDc-XUCBs8CSZI1qGNbUj6CpUMyA8suDUpnZ1WO3lHEUFuk1lk3s4wJtM)